chore: remove legacy CLI from package
