+++ .github/workflows/ci.yml
@@ -1,30 +1,14 @@
-name: CI
-on:
-  push:
-    branches: [main]
-  pull_request:
+name: test
+on: [push, pull_request]
 
 jobs:
-  test:
+  smoke:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: '3.11'
-      - name: Install project deps
-        run: |
-          python -m pip install -e .[ci]
-          python -m pip install pytest
-          python -m pip install -r herg-agent/requirements.txt
-
-          # sanity print
-          python - <<'PY'
-          import sys, numpy, yaml, IPython
-          print("NumPy", numpy.__version__, "PyYAML", yaml.__version__, "IPython", IPython.__version__, file=sys.stderr)
-          PY
-      - name: Run unit tests
-        run: pytest -q
-      - name: Run style checks
-        run: python tools/run_precommit.py
+      - uses: actions/setup-python@v5
+        with: {python-version: '3.11'}
+      - run: |
+          python -m pip install -e .[dev]
+      - run: herg-run --demo text
+      - run: pytest -q
