+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy==1.26.4
+faiss-cpu==1.7.4
+fastapi==0.110.0
+uvicorn==0.29.0
+orjson==3.9.15
+httpx==0.27.0
