--- setup.cfg
@@ -1,15 +1,0 @@
-[metadata]
-name = herg
-version = 0.1.0
-description = HERG – Hypervector Engine with Astrocyte-inspired Capsules
-
-[options]
-packages = find:
-python_requires = >=3.8
-
-[options.extras_require]
-ci =
-    numpy==1.26.4
-    PyYAML==6.0.1
-    ipython==8.24.0
-dev = pre-commit>=3.7.0
